+++ kii/kii_impl.h
@@ -4,8 +4,6 @@
 #include "kii.h"
 #include "khc.h"
 
-int _kii_set_content_length(kii_t* kii, size_t content_length);
-
 kii_code_t _convert_code(khc_code khc_c);
 
 void _reset_buff(kii_t* kii);
